==> Report/06 - kNN model.ipynb <==
#Data Pre-precessing

import numpy as np
import pandas as pd
import csv
from sklearn import preprocessing

def word2num(x,y):
    return [i for i in range(len(y)) if y[i] == x]

def data_precessing(source_datafile,handled_datafile):
    file_data = open(handled_datafile,'w',newline='') #Write dataset

    with open(source_datafile,'r') as data_soure:
        csv_read = csv.reader(data_soure)
        csv_write = csv.writer(file_data)
        for row in csv_read:
            temp = np.array(row)
            
            #Protocol
            protocol_list = ['tcp','udp','icmp']
            if temp[1] in protocol_list:
                temp[1] = word2num(temp[1],protocol_list)[0]

            #Service
            service_list = ['aol','auth','bgp','courier','csnet_ns','ctf','daytime','discard','domain','domain_u',
                    'echo','eco_i','ecr_i','efs','exec','finger','ftp','ftp_data','gopher','harvest','hostnames',
                    'http','http_2784','http_443','http_8001','imap4','IRC','iso_tsap','klogin','kshell','ldap',
                    'link','login','mtp','name','netbios_dgm','netbios_ns','netbios_ssn','netstat','nnsp','nntp',
                    'ntp_u','other','pm_dump','pop_2','pop_3','printer','private','red_i','remote_job','rje','shell',
                    'smtp','sql_net','ssh','sunrpc','supdup','systat','telnet','tftp_u','tim_i','time','urh_i','urp_i',
                    'uucp','uucp_path','vmnet','whois','X11','Z39_50']
            if temp[2] in service_list:
                temp[2] = word2num(temp[2],service_list)[0]
        
            #Flag
            flag_list = ['OTH','REJ','RSTO','RSTOS0','RSTR','S0','S1','S2','S3','SF','SH']
            if temp[3] in flag_list:
                temp[3] = word2num(temp[3],flag_list)[0]

            #State
            if(temp[41] == 'normal'):
                temp[41] = 0
            elif (temp[41] == 'DDoS'):
                temp[41] = 1
            elif (temp[41] == 'UR2'):
                temp[41] = 2
            elif (temp[41] == 'R2L'):
                temp[41] = 3
            else:
                temp[41] = 4
            csv_write.writerow(temp)
        file_data.close()

    return 0
    
def onehot(ori_data, handled_datafile):
    #One-Hot Encoding
    with open('/Users/ziyuanjiang/dstoolbox/set/kddcup_pre') as onehot_file:
        data_matrix = np.loadtxt(onehot_file, delimiter=",")

    with open(ori_data) as onehot_file1:
        data_matrix_source = np.loadtxt(onehot_file1, delimiter=",")

    enc = preprocessing.OneHotEncoder()
    enc.fit(data_matrix[:,[1,2,3]])
    en_matrix = np.zeros((len(data_matrix_source),80))

    for i in range(len(data_matrix_source)):
        en_temp = enc.transform([data_matrix_source[i,[1,2,3]]]).toarray()
        en_matrix[i,:] = en_temp

    
    data_fin = np.hstack((en_matrix,np.delete(data_matrix_source,[1,2,3],axis=1)))
    np.savetxt(handled_datafile,data_fin.astype(int),delimiter=',',fmt='%.0f')
    

#Data Pre-precessing
data_precessing('/Users/ziyuanjiang/dstoolbox/set/train','/Users/ziyuanjiang/dstoolbox/set/train_pre')

data_precessing('/Users/ziyuanjiang/dstoolbox/set/kddcup','/Users/ziyuanjiang/dstoolbox/set/kddcup_pre')
onehot('/Users/ziyuanjiang/dstoolbox/set/train_pre','/Users/ziyuanjiang/dstoolbox/set/train_onehot')

data_precessing('/Users/ziyuanjiang/dstoolbox/set/test','/Users/ziyuanjiang/dstoolbox/set/test_pre')
onehot('/Users/ziyuanjiang/dstoolbox/set/test_pre','/Users/ziyuanjiang/dstoolbox/set/test_onehot')

data_precessing('/Users/ziyuanjiang/dstoolbox/set/labeled','/Users/ziyuanjiang/dstoolbox/set/labeled_pre')
onehot('/Users/ziyuanjiang/dstoolbox/set/labeled_pre','/Users/ziyuanjiang/dstoolbox/set/labeled_onehot')


from sklearn import preprocessing
import pandas as pd

file = '/Users/ziyuanjiang/dstoolbox/set/train_onehot'

with open(file) as file_op:
    data_matrix = np.loadtxt(file_op, delimiter=",")

file = '/Users/ziyuanjiang/dstoolbox/set/test_onehot'

with open(file) as file_op:
    test_data = np.loadtxt(file_op, delimiter=",")


#Normalisation of the dataset and partitioning of the training and test sets
from sklearn import preprocessing
import pandas as pd
from sklearn.model_selection import train_test_split
min_max_scaler = preprocessing.MinMaxScaler()
x_minmax = min_max_scaler.fit_transform(data_matrix[:,0:118])
fin_data = np.hstack((x_minmax,data_matrix[:,118:119]))

x_minmax = min_max_scaler.fit_transform(test_data[:,0:118])
fintest_data = np.hstack((x_minmax,test_data[:,118:119]))


train, test = train_test_split(fin_data, test_size=0.3)

with open('/Users/ziyuanjiang/dstoolbox/set/labeled_onehot') as file_op:
    data_matrix = np.loadtxt(file_op, delimiter=",")

x_minmax = min_max_scaler.fit_transform(data_matrix[:,0:118])
fintest_data = np.hstack((x_minmax,data_matrix[:,118:119]))

en_matrix = np.zeros((len(train),5))

for i in range(len(train)):
    en_matrix[i,int(train[i,118:119])] = 1


#Training models
import matplotlib.pyplot as plt
import numpy as np
from sklearn.datasets import make_blobs
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import train_test_split
from sklearn import neighbors

X = train[:,0:118]
Y = en_matrix
data = make_blobs(n_samples=311232, n_features=118, centers=5,cluster_std=1.0, random_state=8)  

clf = neighbors.KNeighborsClassifier()
clf.fit(X, Y)



matrix = pd.read_csv('/Users/ziyuanjiang/dstoolbox/set/labeled_test', delimiter=' ', )
del matrix['normal']
matrix.to_csv('/Users/ziyuanjiang/dstoolbox/set/labeled',encoding="utf_8", index=None, header=None)
res_test = clf.predict(fintest_data[:,0:118])

res = np.zeros((len(result),1))
for i in range(len(result)):
    for j in range(5):
        if result[i,j] == 1:
            res[i,0] = j


res = np.zeros((len(res_test),1))
for i in range(len(res_test)):
    for j in range(5):
        if res_test[i,j] == 1:
            res[i,0] = j

from sklearn import metrics
print (sum(res==fintest_data[:,118:119])) 
print(metrics.classification_report(res, fintest_data[:,118:119])) 

|        | normal | probe | ur2 | ddos   | r2l |
| ------ | ------ | ----- | --- | ------ | --- |
| normal | 83286  | 104   | 91  | 210    | 624 |
| probe  | 3012   | 270   | 5   | 21     | 50  |
| ur2    | 2453   | 3     | 0   | 0      | 15  |
| ddos   | 7396   | 502   | 6   | 221933 | 16  |
| r2l    | 10045  | 0     | 52  | 5      | 52  |

from sklearn import metrics
print (sum(res==test_data[:,118:119])) 
print(metrics.classification_report(res, test_data[:,118:119])) 

|        | normal | probe | ur2 | ddos  | r2l |
| ------ | ------ | ----- | --- | ----- | --- |
| normal | 3580   | 75    | 281 | 0     | 156 |
| probe  | 598    | 240   | 1   | 4     | 10  |
| ur2    | 356    | 1     | 0   | 0     | 3   |
| ddos   | 1246   | 157   | 6   | 38101 | 3   |
| r2l    | 2015   | 0     | 16  | 1     | 15  |
