==> Report/08-conclusion.ipynb <==
# Conclusion

This assessment aims to detect the anormaly network connections and predict the attack types.This document summarises the assessment in three parts, the selection and division of the data, the model fitting for binary and multiclass classification and the performance evaluation.

## Selection and Division of the Data

As mentioned in the introductory document, due to our limited computer processing capacity, we used kddcup 10% of the data. And due to the variety of attack types, we have divided the 22 types of attacks into four main categories.

For the division of the data, we first divided the 10% labeled dataset, with part of it serving as the training set and the remainder as the test set. We also used the 'labeled' dataset provided by kddcup1999 as the test set, which contains a dozen attack types that were not included in the training set.

## Binary Classification

In the binary classification model, we used logistic regression, decision tree, xgboost and kNN as classifiers. All four models have a relatively high accuracy rate for classification.

## Multivariate Classification

Three models were used in the multi-classification model, Xgboost, Multi- logistic Regression and kNN. the accuracy and comparison of these three types of models can be found in document 07.
